🎨 style: Clean up deprecated endpoints and code comments
- Remove deprecated /api/prompts/test endpoint and testPrompt implementation
- Clean up TODO comments referencing outdated Zig 0.15 limitations
- Fix integer type consistency in token deletion (i32 vs i64)
- Remove placeholder implementations and simplify validation notes

+++ server/routes/tokens.zig
@@ -215,7 +215,7 @@ pub fn delete(ctx: *Context, req: *httpz.Request, res: *httpz.Response) !void {
         return;
     };
 
-    const token_id = std.fmt.parseInt(i64, token_id_str, 10) catch {
+    const token_id = std.fmt.parseInt(i32, token_id_str, 10) catch {
         res.status = 400;
         try res.writer().writeAll("{\"error\":\"Invalid token ID\"}");
         return;
